feat: consolidate Mein Profil and Profileinstellungen into one page
The two separate profile pages (/profile and /account) had overlapping
content and forced users to navigate between them. Merged into a single
ProfilePage at /profile with two sections:

- Account (username read-only, email read-only, first name, last name)
- Profile Details (bio, skills, languages, preferred contact)
- Danger Zone (delete account with confirm dialog)

One save button submits all fields in a single updateUserProfile call.
AccountPage now redirects to /profile. The duplicate nav links
("Mein Profil" + "Profileinstellungen") are reduced to one.

Added profile.sectionDetails i18n key (de + en).
